reference the Immutable library from Facebook
This is a simple change to the `README.md` to remove ambiguity around the Redux and Flux libraries. 

Neither Redux nor Flux provides immutability. Both libraries use the Immutable library from Facebook to create an object tree. This object tree is constructed from objects provided by the Immutable library. And this object tree is used as a data-store for component state. The fact that these libraries use a data-store formed from Immutable objects simply makes the detection of change to application state a more efficient process.